+++ src/Language/Python/Internal/Syntax.hs
@@ -14,7 +14,6 @@ module Language.Python.Internal.Syntax
   , module Language.Python.Internal.Syntax.AugAssign
   , module Language.Python.Internal.Syntax.CommaSep
   , module Language.Python.Internal.Syntax.Comment
-  , module Language.Python.Internal.Syntax.Expr
   , module Language.Python.Internal.Syntax.Ident
   , module Language.Python.Internal.Syntax.Import
   , module Language.Python.Internal.Syntax.Module
@@ -35,7 +34,6 @@ import Text.Parser.Token.Highlight (Highlight(..))
 import Language.Python.Internal.Syntax.AugAssign
 import Language.Python.Internal.Syntax.CommaSep
 import Language.Python.Internal.Syntax.Comment
-import Language.Python.Internal.Syntax.Expr
 import Language.Python.Internal.Syntax.Ident
 import Language.Python.Internal.Syntax.Import
 import Language.Python.Internal.Syntax.Module
